+++ src/main/java/com/jyp/justplan/domain/plan/application/PlanService.java
@@ -3,15 +3,12 @@ package com.jyp.justplan.domain.plan.application;
 import com.jyp.justplan.domain.city.domain.City;
 import com.jyp.justplan.domain.city.domain.CityRepository;
 import com.jyp.justplan.domain.city.dto.response.CityResponse;
-import com.jyp.justplan.domain.mbti.domain.Mbti;
 import com.jyp.justplan.domain.plan.application.budget.BudgetService;
 import com.jyp.justplan.domain.plan.application.expense.ExpenseService;
 import com.jyp.justplan.domain.plan.application.tag.PlanTagService;
 import com.jyp.justplan.domain.plan.domain.Plan;
 import com.jyp.justplan.domain.plan.domain.PlanRepository;
 import com.jyp.justplan.domain.plan.domain.UserPlan;
-import com.jyp.justplan.domain.plan.domain.scrap.Scrap;
-import com.jyp.justplan.domain.plan.domain.scrap.ScrapStore;
 import com.jyp.justplan.domain.plan.domain.tag.PlanTag;
 import com.jyp.justplan.domain.plan.dto.request.PlanIdRequest;
 import com.jyp.justplan.domain.plan.dto.request.PlanCreateRequest;
